Remove redundant sentence
This section is about how to re-fill missing seats. How to vote in the
presence of missing seats is a separate question, and already covered
above in the sections no how SC voting works.

Restating an existing rule "denormalizes" the constitution, and risks
the accidental creation of a contradiction. It also reduces clarity as
to which information is to be found in which section.