Level- not edge- trigger for special special elections
Right now the language is written in an "edge trigger" manner (i.e. "if
a seat becomes vacant..."). But I believe the intent is in fact a "level
trigger" (e.g. "when some seats are vacant....").

This switches the language to level triggering.
As a consequence, it is also explicit that at the time the SC decides to
hold a special election, more than one seat might be vacant.